default.nix: Install only executables by default
Problem: derivation not filtered by `justStaticExecutables` contains all
kinds of libraries that are not needed when installing with `nix-env`,
but they are the same for all Haskell expressoins and thus result in
file-name conflicts.

Solution: Add an optional parameter that controls whether we filter the
derivation or not, set it to “filter” by default, but override when
importing from other files, e.g. `shell.nix`.