Indent expressions properly
The main change is that the indentation level inside nests is now
relative to the indentation of the line it starts on, rather than
relative to the indentation that the next line would have started on.

The remaining changes are mostly related to expressions with closers
that should not be indented relative to the start of the expression,
like sets, lists and indented strings. This can now be easily specified
by nesting them with zero spaces, forcing all parts including the closer
to be at the same indentation as the opener, except for the body which
is nested again with two spaces.

This commit solves the problem for most of the common cases, but not for
parenthesized expressions for example. The remaining indentation issues
should be solved one by one because they require a little more
consideration.