Deployment: run `aws cp` before `aws sync` to replace existing files
`aws s3 sync` compares file size and timestamps to decide if a file in
the bucket needs to be replaced, but all files in nix store have
timestamp=1. So if a new file happens to have the same size as the old
one, `aws s3 sync` would decide that it does not need to be updated.
Run `aws cp` before `aws sync` to update existing files unconditionally