Better trailing comment handling
If a supposedly trailing comment is directly followed by line comments of same indentation,
treat it as part of them instead of as a trailing comment.

This allows us to better distinguish between "comment on the entire set" and "comment on the first element".
Now we can safely move trailing comments on `{` and `[` up before the token, without fear of wrongly breaking up sentences.