Improve priority group handling
It now also takes into account whether post fits onto a single line or not.
Previously this did not matter, as all instances only had whitespace in there,
so force-unexpanding worked as a hack. But now it is time to fully implement the feature.

No big changes to the layouting algorithm itself were made, however it needed to be
refactored such that the interface it provides only processes one item at a time.
(The algorithm already did that, it's just that the function always also processed the rest.)
This has been done by making use of the State monad.

This should cause no changes to the output format.