Fix idempotency, enable --verify in tests
The basic issue is that we must not emit a trailing comment at +2 of the current indentation.
This is because, that comment will then be parsed as a leading comment to the next line,
instead of a trailing comment to the current line, leading to a different output.

Input:

```nix
foo bar baz ( # comment
  b
)
```

Output 1:

```nix
foo
  bar
  baz
  ( # comment
    b
  )
```

Output 2:

```nix
foo
  bar
  baz
  (
    # comment
    b
  )
```

Simply guaranteeing that all comments will be moved up such that such an emission never happens is not feasible,
due to a large number of edge cases.

Instead, the currently implemented solution is to track trailing comments specially in the intermediate representation,
and let the renderer handle this. The render will check for such a situation, and insert an additional space to avoid it:

```nix
foo
  bar
  baz
  (  # comment
    b
  )
```

This may look ugly, but in real-world code it should be sufficiently rare. Most of the time, moving comments up already
happens and thus eliminates a lot of potential cases entirely. This is merely a fallback. Most of the time, the output
code will look like this instead:

```nix
foo
  bar
  baz
  # comment
  (b)
```

The double space before a trailing comment is most likely to occur with single-letter identifiers.