mv .envrc → .envrc.example
It isn’t a good idea to make a personal configuration file hard to work
with since it’s being tracked by the repository. This will a) trigger
``direnv`` to fire off its allow script before a use gets a chance to
see what arbitrary code is about to be executed on their behalf & b) make
making modifications more difficult as it would be easy to accidentally
commit those modifications *or* requires addings all sorts of
unnecessary extra configuration to allow a safe place to put in custom
environment variables.

The simple solution is to just leave an example for those that want it.