Track the source line position in Ann
This is finally the new field of Ann that past commits have been working
up to.

Because it's not an optional field, the ann function is extended to
take an extra argument for its initialisation, which then automatically
prompts the connection of source locations from the original to the newly
created value.