Always preserve expanded lists
This changes the formatting to take into account whether lists
are expanded onto multiple lines already, and if so, preserves that.

This is desirable because if a user writes an list over
multiple lines, there's most likely an intention to add more elements,
which would become harder when formatted.

An example of this is build inputs, which usually start out with just a
single element, but usually get more elements later. This would've been
contracted onto a single line before this change:

    buildInputs = [
      libfoo
    ];

This change conforms to the standard due to this line:

> The formatter may take the input formatting into account in some cases in order to preserve multi-line syntax elements (which would otherwise have been contracted by the rules).