Always preserve expanded attribute parameters
This changes the formatting to take into account whether attribute
parameters are expanded onto multiple lines already,
and if so, preserves that.

This is desirable because if a user writes an attribute parameter over
multiple lines, there's most likely an intention to add more attributes,
which would become harder when formatted.

An example of this is package recipes, which usually start out with just a
single attribute argument, but usually get more later. This would've been
contracted onto a single line before this change:

    {
      stdenv,
    }:

This change conforms to the standard due to this line:

> The formatter may take the input formatting into account in some cases in order to preserve multi-line syntax elements (which would otherwise have been contracted by the rules).