Make empty inherit's more consistent and convenient
This might happen if you temporarily remove all inherits, in which case
the formatter shouldn't remove the spacing that was there.

E.g. I might have this initially

    inherit a b;

    inherit
      c
      d
      ;

removing all the bindings gives me

    inherit ; # I already have the space that I'll need

    inherit # I can just press enter here to add one
      ;

which shouldn't be changed, because it's easier like this to insert new
entries again, compared to

    inherit;

This also makes it more consistent with the parent commit's change