Create .pre-commit-hooks.yaml
This will allow people to enforce the style of .nix files in their own
repos via the pre-commit tool. This change is related to #221, but
it doesn’t actually implement the idea proposed in that issue.

When writing the instructions for how to use the nixfmt pre-commit hook,
I tried to make it clear that there’s a difference between the
pre-commit tool and Git pre-commit hooks in general. I tried to make the
difference between those two things clear for two reasons:

1. Both of those things are called pre-commit which makes it easy for
people to mix them up [1].

2. Depending on how #221 shakes out, the nixfmt project might end up
adding another alternative mechanism for using nixfmt as part of a Git
pre-commit hook script. I’m preemptively using language that makes it
easy to differentiate between those two different mechanisms just in
case there ends up being two different mechanisms in the future.

[1]: <https://github.com/DescentDevelopers/Descent3/pull/450#issuecomment-2184023652>