Fix merge mix-up in mergetool
Turns out, `git merge-tool` wants the base version in the _second_
argument, and not the first! This was never noticed, because for changes
_only_ related to formatting, it won't matter, since for usual rebases
and merges, the actual first argument, the local version, only contains
the formatting changes, which can be entirely ignored anyways.

So this only causes problems if there's not exclusively
formatting-related changes, which can of course happen.